+++ samples/UseWords.hs
@@ -17,10 +17,7 @@ import Paths
 main :: IO ()
 main = do
     static <- getStaticDir
-    startGUI defaultConfig
-        { tpPort       = 10000
-        , tpStatic     = Just static
-        } setup
+    startGUI defaultConfig { tpStatic = Just static } setup
 
 
 setup :: Window -> UI ()