+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Dependency on `text` package now from version 0.11 to 1.1.*.
 * Dependency on `aeson` package replaces the former dependency on the `json` package.
 * Unicode characters are now transmitted correctly to the browser. #75, #62.
+* Change default port number to 8023. #64
 
 **0.4.0.2** -- Bugfix release.